Minor: remove warnings and refactoring

+++ compiler/frontend/src/org/jetbrains/kotlin/parsing/KotlinExpressionParsing.java
@@ -1070,16 +1070,14 @@ public class KotlinExpressionParsing extends AbstractKotlinParsing {
             }
         }
 
-        if (!paramsFound) {
-            if (preferBlock) {
-                literal.drop();
-                parseStatements();
-                expect(RBRACE, "Expecting '}'");
-                literalExpression.done(BLOCK);
-                myBuilder.restoreNewlinesState();
+        if (!paramsFound && preferBlock) {
+            literal.drop();
+            parseStatements();
+            expect(RBRACE, "Expecting '}'");
+            literalExpression.done(BLOCK);
+            myBuilder.restoreNewlinesState();
 
-                return;
-            }
+            return;
         }
 
         PsiBuilder.Marker body = mark();
